Close #954

+++ lib/Doctrine/ORM/Cache/Region/DefaultRegion.php
@@ -20,11 +20,13 @@
 
 namespace Doctrine\ORM\Cache\Region;
 
+use Doctrine\Common\Cache\Cache as CacheAdapter;
+use Doctrine\Common\Cache\ClearableCache;
+use Doctrine\ORM\Cache\CacheEntry;
+use Doctrine\ORM\Cache\CacheKey;
+use Doctrine\ORM\Cache\CollectionCacheEntry;
 use Doctrine\ORM\Cache\Lock;
 use Doctrine\ORM\Cache\Region;
-use Doctrine\ORM\Cache\CacheKey;
-use Doctrine\ORM\Cache\CacheEntry;
-use Doctrine\Common\Cache\CacheProvider;
 
 /**
  * The simplest cache region compatible with all doctrine-cache drivers.
@@ -35,7 +37,7 @@ use Doctrine\Common\Cache\CacheProvider;
 class DefaultRegion implements Region
 {
     /**
-     * @var \Doctrine\Common\Cache\CacheProvider
+     * @var CacheAdapter
      */
     protected $cache;
 
@@ -50,11 +52,11 @@ class DefaultRegion implements Region
     protected $lifetime = 0;
 
     /**
-     * @param string                                $name
-     * @param \Doctrine\Common\Cache\CacheProvider  $cache
-     * @param integer                               $lifetime
+     * @param string       $name
+     * @param CacheAdapter $cache
+     * @param integer      $lifetime
      */
-    public function __construct($name, CacheProvider $cache, $lifetime = 0)
+    public function __construct($name, CacheAdapter $cache, $lifetime = 0)
     {
         $this->cache    = $cache;
         $this->name     = (string) $name;
@@ -96,6 +98,37 @@ class DefaultRegion implements Region
     /**
      * {@inheritdoc}
      */
+    public function getMultiple(CollectionCacheEntry $collection)
+    {
+        $keysToRetrieve = array();
+
+        foreach ($collection->identifiers as $index => $key) {
+            $keysToRetrieve[$index] = $this->name . '_' . $key->hash;
+        }
+
+        $items = array_filter(
+            array_map([$this->cache, 'fetch'], $keysToRetrieve),
+            function ($retrieved) {
+                return false !== $retrieved;
+            }
+        );
+
+        if (count($items) !== count($keysToRetrieve)) {
+            return null;
+        }
+
+        $returnableItems = array();
+
+        foreach ($keysToRetrieve as $index => $key) {
+            $returnableItems[$index] = $items[$key];
+        }
+
+        return $returnableItems;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function put(CacheKey $key, CacheEntry $entry, Lock $lock = null)
     {
         return $this->cache->save($this->name . '_' . $key->hash, $entry, $this->lifetime);
@@ -114,6 +147,13 @@ class DefaultRegion implements Region
      */
     public function evictAll()
     {
+        if (! $this->cache instanceof ClearableCache) {
+            throw new \BadMethodCallException(sprintf(
+                'Clearing all cache entries is not supported by the supplied cache adapter of type %s',
+                get_class($this->cache)
+            ));
+        }
+
         return $this->cache->deleteAll();
     }
 }
